+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Pocketmine/Model/PocketmineProject.cs
@@ -1,0 +1,10 @@
+﻿using MinecraftJars.Core.Projects;
+
+namespace MinecraftJars.Plugin.Pocketmine.Model;
+
+public record PocketmineProject(
+    Group Group,
+    string Name,
+    string Description,
+    string Url,
+    byte[] Logo) : IProject;