+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Pocketmine/Model/PocketmineVersion.cs
@@ -1,0 +1,15 @@
+﻿using MinecraftJars.Core.Downloads;
+using MinecraftJars.Core.Projects;
+using MinecraftJars.Core.Versions;
+
+namespace MinecraftJars.Plugin.Pocketmine.Model;
+
+public record PocketmineVersion(
+    IProject Project,
+    string Version) : IVersion
+{
+    internal PocketmineDownload Download { get; init; } = default!;
+    
+    public Task<IDownload> GetDownload(DownloadOptions? options = null, CancellationToken cancellationToken = default!) => 
+        PocketmineVersionFactory.GetDownload(options ?? new DownloadOptions(), this, cancellationToken);
+}